Fixes for dropping columns if they do not exist yet

+++ database/migrations/2022_11_06_215526_drop_html_backups_column_from_backups_table.php
@@ -13,9 +13,17 @@ return new class extends Migration
      */
     public function up()
     {
+
+        if (Schema::hasColumn('backups', 'html_backup'))
+        {
+            Schema::table('backups', function (Blueprint $table)
+            {
+                $table->dropColumn('html_backup');
+            });
+        }
+
         Schema::table('backups', function (Blueprint $table) {
             $table->string('disk')->nullable();
-            $table->dropColumn('html_backup');
         });
     }
